Nest xenophile's WIT model under the Wit namespace
`WitFunction`, `WitItem`, `WitInterface`, `WitWorldModel`, `WitDocument`,
`WitParseError` and `WitParser` become members of `object Wit`.

The proposal's row for this file named `WitDeclaration → Wit.Declaration`, but no
type of that name exists anywhere in the tree: the file is a container named after
a type it does not declare, and the six model types above are its real contents.

Nesting surfaced a collision the "check the outer object for a member of the
target name" rule does not catch, because the clash is one level deeper: `enum
Item` has a case `Function`, so the newly-nested `Wit.Function` became ambiguous
with `Wit.Item.Function` inside the enum's own body. Both the case's parameter
types and the pattern that matches it are now qualified.

`WitDialect` is untouched (reflective lookup), and `WitCase`/`WitError`/
`WitHandle`/`WitVariant` stay put: they live in the downstream `xenophile.wasm`
component, which `Wit.scala` cannot see. Verified with `xenophile.test.run`
(203 passed).
